requirements/cpu.txt
-i https://pypi.org/simple

annotated-types==0.5.0 ; python_version >= '3.7'

anyio==3.7.1 ; python_version >= '3.7'

certifi==2023.7.22 ; python_version >= '3.6'

charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'

click==8.1.7 ; python_version >= '3.7'

colorama==0.4.6 ; platform_system == 'Windows'

exceptiongroup==1.1.3 ; python_version < '3.11'

fastapi==0.103.1

filelock==3.12.4 ; python_version >= '3.8'

fsspec==2023.9.2 ; python_version >= '3.8'

h11==0.14.0 ; python_version >= '3.7'

huggingface-hub==0.17.2 ; python_full_version >= '3.8.0'

idna==3.4 ; python_version >= '3.5'

jinja2==3.1.2 ; python_version >= '3.7'

markupsafe==2.1.3 ; python_version >= '3.7'

mpmath==1.3.0

networkx==3.1 ; python_version >= '3.8'

numpy==1.26.0

opencv-python==4.8.0.76

packaging==23.1 ; python_version >= '3.7'

pillow==10.0.1

pydantic==2.4.0 ; python_version >= '3.7'

pydantic-core==2.10.0 ; python_version >= '3.7'

python-multipart==0.0.6

pyyaml==6.0.1 ; python_version >= '3.6'

requests==2.31.0 ; python_version >= '3.7'

safetensors==0.3.3

scipy==1.11.2

sniffio==1.3.0 ; python_version >= '3.7'

starlette==0.27.0 ; python_version >= '3.7'

sympy==1.12 ; python_version >= '3.8'

timm==0.9.7

torch==2.0.1

torchvision==0.15.2

tqdm==4.66.1 ; python_version >= '3.7'

typing-extensions==4.8.0 ; python_version >= '3.8'

urllib3==2.0.5 ; python_version >= '3.7'

uvicorn==0.23.2

websockets==11.0.2

